add HTTP::Throwable exceptions
So when we get back Server Errors it would be nice to get exceptions
for those errors and trigger calls to `on_error` rather than letting
them slide past silently.

This commit adds HTTP::Throwable to the mix so when we get an error
response that is translated into the appropriate HTTP::Throwable
exception object. I've subclassed HTTP::Throwable::Factory so that
every exception object can hold a reference to the response object
that generated the exception.