drop back to just `is_error`
We really don't care if it's a `4xx` or a `5xx` error, we're gonna
want to handle them them the same anyway. Additionally if it's a
`3xx` response we don't want to drop it on the floor.

Re-arrange things so that we handle the exceptions but anything else
ends up passed back to the user.