fix it to show HTTP::Throwable::Role::Status roles
So the release that added `HTTP::Throwable` didn't quite work as
fully expected. While it would throw an exception properly that
exception wasn't typed with one of the pre-built ident Roles that
HTTP::Throwable ships with. This is because we were using the generic
constructor rather than the one that applies ident roles.

This patch fixes it by inlining a patch to `HTTP::Throwable::Factory`
that we've submitted upstream, and checking to see if we have an
exception code with an known ident mapping. If we do then we throw a
properly typed exception otherwise we fall back to throwing a generic
exception.